requirements.txt
pillow==10.3.0
streamlit==1.35.0
numpy==1.26.4
tensorflow==2.8.0rc0
pandas==2.1.4
fastbook==0.0.29
imbalanced-learn==0.11.0
matplotlib==3.8.0
scikit-learn==1.2.2